requirements.txt
streamlit
qrcode[pil]
pandas
openpyxl
pillow
reportlab
pymupdf
